week2_alerts/requirements.txt
PyYAML==6.0.2
requests==2.32.3
pytz==2024.1
schedule==1.2.1
python-dotenv==1.0.1
